requirements.txt
matplotlib
neptune-client
numpy
optuna
pandas
Pillow
scikit-image
string_utils
torch
torchvision